iOS: use core image formatter
