+++ ios/Features/WalletTab/Sources/ViewModels/AssetListItemViewModel.swift
@@ -1,6 +1,7 @@
 // Copyright (c). Gem Wallet. All rights reserved.
 
 import Components
+import GemstonePrimitives
 import Primitives
 
 struct AssetListItemViewModel: Identifiable {
