--- ios/Packages/Primitives/Sources/AssetImageFormatter.swift
@@ -1,30 +1,0 @@
-// Copyright (c). Gem Wallet. All rights reserved.
-
-import Foundation
-
-public struct AssetImageFormatter: Sendable {
-    public static let shared = AssetImageFormatter()
-
-    public init() {}
-
-    public func getURL(for assetId: AssetId) -> URL {
-        let urlString = if let tokenId = assetId.tokenId {
-            "\(Constants.assetsURL.absoluteString)/blockchains/\(assetId.chain.rawValue)/assets/\(tokenId)/logo.png"
-        } else {
-            "\(Constants.assetsURL.absoluteString)/blockchains/\(assetId.chain.rawValue)/logo.png"
-        }
-        return URL(string: urlString)!
-    }
-
-    public func getNFTUrl(for assetId: String) -> URL {
-        URL(string: "\(Constants.assetsURL.absoluteString)/nft/assets/\(assetId)/preview")!
-    }
-
-    public func getValidatorUrl(chain: Chain, id: String) -> URL {
-        URL(string: "\(Constants.assetsURL.absoluteString)/blockchains/\(chain.rawValue)/validators/\(id)/logo.png")!
-    }
-
-    public func getListUrl(for id: String) -> URL {
-        URL(string: "\(Constants.assetsURL.absoluteString)/lists/\(id).png")!
-    }
-}
